* Fix MarshalSize::process for BufferCarrier since we can't serialize   a size_t in a cross-platform way

+++ serialize/MarshalSerialize.cc
@@ -471,7 +471,7 @@ MarshalSize::process(const char*            name,
 {
     (void) carrier;
 
-    size_t size;
+    u_int32_t size;
     process(name, &size);
     size_ += carrier->len();
 }
